pyproject.toml
[tool.poetry]
authors = ["William Sampaio <1997wbms@gmail.com>"]
description = "Implementação em python do layout padrão de arrecadação/recebimento com utilização do código de barras da Febraban."
license = "MIT"
name = "febraban-barcode"
packages = [{include = "febraban_barcode"}]
readme = "README.md"
version = "0.3.0"

[tool.poetry.urls]
"Código" = "https://github.com/WilliamSampaio/febraban_barcode"
"Documentação" = "https://febraban-barcode.readthedocs.io/"

[tool.poetry.dependencies]
pillow = "^10.0.0"
python = "^3.11"
python-barcode = "^0.15.1"

[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
isort = "^5.12.0"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
tinydb = "^4.8.0"

[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.1.21"
mkdocstrings = "^0.22.0"
mkdocstrings-python = "^1.2.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.taskipy.tasks]
docs = "mkdocs serve"
lint = "isort . && blue ."
test = "pytest -x -v --ff --cov=febraban_barcode --cov-report=html tests/"
